change model to use Decimal instead of float for amount and balance

+++ casa/models.py
@@ -1,5 +1,6 @@
 import enum
 from datetime import datetime
+from decimal import Decimal
 from typing import List
 
 from sqlalchemy import (
@@ -35,8 +36,8 @@ class Account(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     account_num: Mapped[str] = mapped_column(String(32), unique=False)
     currency: Mapped[str] = mapped_column(String(3))
-    balance: Mapped[float] = mapped_column(DECIMAL(14, 2))
-    avail_balance: Mapped[float] = mapped_column(DECIMAL(14, 2))
+    balance: Mapped[Decimal] = mapped_column(DECIMAL(14, 2))
+    avail_balance: Mapped[Decimal] = mapped_column(DECIMAL(14, 2))
     status: Mapped[StatusEnum] = mapped_column(Enum(StatusEnum), default=StatusEnum.ACTIVE)
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
 
@@ -52,7 +53,7 @@ class Transaction(Base):
     ref_id: Mapped[str] = mapped_column(String(32), index=True)
     trx_date: Mapped[str] = mapped_column(String(10))
     currency: Mapped[str] = mapped_column(String(3))
-    amount: Mapped[float] = mapped_column(DECIMAL(14, 2))
+    amount: Mapped[Decimal] = mapped_column(DECIMAL(14, 2))
     memo: Mapped[str] = mapped_column(String(100))
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
 
@@ -69,7 +70,7 @@ class Transfer(Base):
     ref_id: Mapped[str] = mapped_column(String(32), unique=True)
     trx_date: Mapped[str] = mapped_column(String(10), index=True)
     currency: Mapped[str] = mapped_column(String(3))
-    amount: Mapped[float] = mapped_column(DECIMAL(14, 2))
+    amount: Mapped[Decimal] = mapped_column(DECIMAL(14, 2))
     memo: Mapped[str] = mapped_column(String(100))
     debit_account_num: Mapped[str] = mapped_column(String(32))
     credit_account_num: Mapped[str] = mapped_column(String(32))
